+++ lib/elixir_sense/core/compiler.ex
@@ -1,5 +1,5 @@
 defmodule ElixirSense.Core.Compiler do
-  import ElixirSense.Core.State, except: [expand: 2, expand: 3, no_alias_expansion: 1]
+  import ElixirSense.Core.State
   alias ElixirSense.Core.State
   require Logger
   alias ElixirSense.Core.Introspection
@@ -2605,2273 +2605,6 @@ defmodule ElixirSense.Core.Compiler do
           []
         end
       end
-    end
-  end
-
-  defmodule Utils do
-    def generated([{:generated, true} | _] = meta), do: meta
-    def generated(meta), do: [{:generated, true} | meta]
-
-    def split_last([]), do: {[], []}
-
-    def split_last(list), do: split_last(list, [])
-
-    defp split_last([h], acc), do: {Enum.reverse(acc), h}
-
-    defp split_last([h | t], acc), do: split_last(t, [h | acc])
-
-    def split_opts(args) do
-      case split_last(args) do
-        {outer_cases, outer_opts} when is_list(outer_opts) ->
-          case split_last(outer_cases) do
-            {inner_cases, inner_opts} when is_list(inner_opts) ->
-              {inner_cases, inner_opts ++ outer_opts}
-
-            _ ->
-              {outer_cases, outer_opts}
-          end
-
-        _ ->
-          {args, []}
-      end
-    end
-
-    def get_line(opts) when is_list(opts) do
-      case Keyword.fetch(opts, :line) do
-        {:ok, line} when is_integer(line) -> line
-        _ -> 0
-      end
-    end
-
-    def extract_guards({:when, _, [left, right]}), do: {left, extract_or_guards(right)}
-    def extract_guards(term), do: {term, []}
-
-    def extract_or_guards({:when, _, [left, right]}), do: [left | extract_or_guards(right)]
-    def extract_or_guards(term), do: [term]
-
-    def select_with_cursor(ast_list) do
-      Enum.find(ast_list, &has_cursor?/1)
-    end
-
-    def has_cursor?(ast) do
-      # TODO rewrite to lazy prewalker
-      {_, result} =
-        Macro.prewalk(ast, false, fn
-          _node, true ->
-            {nil, true}
-
-          {:__cursor__, _, list}, _state when is_list(list) ->
-            {nil, true}
-
-          node, false ->
-            {node, false}
-        end)
-
-      result
-    end
-
-    def defdelegate_each(fun, opts) when is_list(opts) do
-      # TODO Remove on elixir v2.0
-      append_first? = Keyword.get(opts, :append_first, false)
-
-      {name, args} =
-        case fun do
-          {:when, _, [_left, right]} ->
-            raise ArgumentError,
-                  "guards are not allowed in defdelegate/2, got: when #{Macro.to_string(right)}"
-
-          _ ->
-            case Macro.decompose_call(fun) do
-              {_, _} = pair -> pair
-              _ -> raise ArgumentError, "invalid syntax in defdelegate #{Macro.to_string(fun)}"
-            end
-        end
-
-      as = Keyword.get(opts, :as, name)
-      as_args = build_as_args(args, append_first?)
-
-      {name, args, as, as_args}
-    end
-
-    defp build_as_args(args, append_first?) do
-      as_args = :lists.map(&build_as_arg/1, args)
-
-      case append_first? do
-        true -> tl(as_args) ++ [hd(as_args)]
-        false -> as_args
-      end
-    end
-
-    # elixir validates arg
-    defp build_as_arg({:\\, _, [arg, _default_arg]}), do: arg
-    defp build_as_arg(arg), do: arg
-  end
-
-  defmodule Clauses do
-    alias ElixirSense.Core.Compiler, as: ElixirExpand
-    alias ElixirSense.Core.Compiler.Utils, as: ElixirUtils
-    alias ElixirSense.Core.State
-    alias ElixirSense.Core.TypeInference
-
-    def match(fun, expr, after_s, _before_s, %{context: :match} = e) do
-      fun.(expr, after_s, e)
-    end
-
-    def match(fun, expr, after_s, before_s, e) do
-      %{vars: current, unused: unused} = after_s
-      %{vars: {read, _write}, prematch: prematch} = before_s
-
-      call_s = %{
-        before_s
-        | prematch: {read, unused, :none},
-          unused: unused,
-          vars: current,
-          calls: after_s.calls,
-          lines_to_env: after_s.lines_to_env,
-          vars_info: after_s.vars_info,
-          cursor_env: after_s.cursor_env,
-          closest_env: after_s.closest_env
-      }
-
-      call_e = Map.put(e, :context, :match)
-      {e_expr, %{vars: new_current, unused: new_unused} = s_expr, ee} = fun.(expr, call_s, call_e)
-
-      end_s = %{
-        after_s
-        | prematch: prematch,
-          unused: new_unused,
-          vars: new_current,
-          calls: s_expr.calls,
-          lines_to_env: s_expr.lines_to_env,
-          vars_info: s_expr.vars_info,
-          cursor_env: s_expr.cursor_env,
-          closest_env: s_expr.closest_env
-      }
-
-      end_e = Map.put(ee, :context, Map.get(e, :context))
-      {e_expr, end_s, end_e}
-    end
-
-    def clause(fun, {:->, clause_meta, [_, _]} = clause, s, e)
-        when is_function(fun, 4) do
-      clause(fn x, sa, ea -> fun.(clause_meta, x, sa, ea) end, clause, s, e)
-    end
-
-    def clause(fun, {:->, meta, [left, right]}, s, e) do
-      {e_left, sl, el} = fun.(left, s, e)
-      {e_right, sr, er} = ElixirExpand.expand(right, sl, el)
-      {{:->, meta, [e_left, e_right]}, sr, er}
-    end
-
-    def clause(fun, expr, s, e) do
-      # try to recover from error by wrapping the expression in clause
-      # elixir raises here bad_or_missing_clauses
-      clause(fun, {:->, [], [[expr], :ok]}, s, e)
-    end
-
-    def head([{:when, meta, [_ | _] = all}], s, e) do
-      {args, guard} = ElixirUtils.split_last(all)
-      prematch = s.prematch
-
-      {{e_args, e_guard}, sg, eg} =
-        match(
-          fn _ok, sm, em ->
-            {e_args, sa, ea} = ElixirExpand.expand_args(args, sm, em)
-
-            {e_guard, sg, eg} =
-              guard(guard, %{sa | prematch: prematch}, %{ea | context: :guard})
-
-            type_info = Guard.type_information_from_guards(e_guard)
-
-            sg = merge_inferred_types(sg, type_info)
-
-            {{e_args, e_guard}, sg, eg}
-          end,
-          :ok,
-          s,
-          s,
-          e
-        )
-
-      {[{:when, meta, e_args ++ [e_guard]}], sg, eg}
-    end
-
-    def head(args, s, e) do
-      match(&ElixirExpand.expand_args/3, args, s, s, e)
-    end
-
-    def guard({:when, meta, [left, right]}, s, e) do
-      {e_left, sl, el} = guard(left, s, e)
-      {e_right, sr, er} = guard(right, sl, el)
-      {{:when, meta, [e_left, e_right]}, sr, er}
-    end
-
-    def guard(guard, s, e) do
-      {e_guard, sg, eg} = ElixirExpand.expand(guard, s, e)
-      {e_guard, sg, eg}
-    end
-
-    # case
-
-    @valid_case_opts [:do]
-
-    def case(e_expr, [], s, e) do
-      # elixir raises here missing_option
-      # emit a fake do block
-      case(e_expr, [do: []], s, e)
-    end
-
-    def case(_e_expr, opts, s, e) when not is_list(opts) do
-      # elixir raises here invalid_args
-      # there may be cursor
-      ElixirExpand.expand(opts, s, e)
-    end
-
-    def case(e_expr, opts, s, e) do
-      # expand invalid opts in case there's cursor
-      {_ast, s, _e} = ElixirExpand.expand(opts |> Keyword.drop(@valid_case_opts), s, e)
-
-      opts = sanitize_opts(opts, @valid_case_opts)
-
-      match_context = TypeInference.type_of(e_expr, e.context)
-
-      {case_clauses, sa} =
-        Enum.map_reduce(opts, s, fn x, sa ->
-          expand_case(x, match_context, sa, e)
-        end)
-
-      {case_clauses, sa, e}
-    end
-
-    defp expand_case({:do, _} = do_clause, match_context, s, e) do
-      expand_clauses(
-        fn c, s, e ->
-          case head(c, s, e) do
-            {[h | _] = c, s, e} ->
-              clause_vars_with_inferred_types =
-                TypeInference.find_typed_vars(h, match_context, :match)
-
-              s = State.merge_inferred_types(s, clause_vars_with_inferred_types)
-
-              {c, s, e}
-
-            other ->
-              other
-          end
-        end,
-        do_clause,
-        s,
-        e
-      )
-    end
-
-    # cond
-
-    @valid_cond_opts [:do]
-
-    def cond([], s, e) do
-      # elixir raises here missing_option
-      # emit a fake do block
-      cond([do: []], s, e)
-    end
-
-    def cond(opts, s, e) when not is_list(opts) do
-      # elixir raises here invalid_args
-      # there may be cursor
-      ElixirExpand.expand(opts, s, e)
-    end
-
-    def cond(opts, s, e) do
-      # expand invalid opts in case there's cursor
-      {_ast, s, _e} = ElixirExpand.expand(opts |> Keyword.drop(@valid_cond_opts), s, e)
-
-      opts = sanitize_opts(opts, @valid_cond_opts)
-
-      {cond_clauses, sa} =
-        Enum.map_reduce(opts, s, fn x, sa ->
-          expand_cond(x, sa, e)
-        end)
-
-      {cond_clauses, sa, e}
-    end
-
-    defp expand_cond({:do, _} = do_clause, s, e) do
-      expand_clauses(&ElixirExpand.expand_args/3, do_clause, s, e)
-    end
-
-    # receive
-
-    @valid_receive_opts [:do, :after]
-
-    def receive([], s, e) do
-      # elixir raises here missing_option
-      # emit a fake do block
-      receive([do: []], s, e)
-    end
-
-    def receive(opts, s, e) when not is_list(opts) do
-      # elixir raises here invalid_args
-      # there may be cursor
-      ElixirExpand.expand(opts, s, e)
-    end
-
-    def receive(opts, s, e) do
-      # expand invalid opts in case there's cursor
-      {_ast, s, _e} = ElixirExpand.expand(opts |> Keyword.drop(@valid_receive_opts), s, e)
-
-      opts = sanitize_opts(opts, @valid_receive_opts)
-
-      {receive_clauses, sa} =
-        Enum.map_reduce(opts, s, fn x, sa ->
-          expand_receive(x, sa, e)
-        end)
-
-      {receive_clauses, sa, e}
-    end
-
-    defp expand_receive({:do, {:__block__, _, []}} = do_block, s, _e) do
-      {do_block, s}
-    end
-
-    defp expand_receive({:do, _} = do_clause, s, e) do
-      # no point in doing type inference here, we have no idea what message we may get
-      expand_clauses(&head/3, do_clause, s, e)
-    end
-
-    defp expand_receive({:after, [_ | _]} = after_clause, s, e) do
-      expand_clauses(&ElixirExpand.expand_args/3, after_clause, s, e)
-    end
-
-    defp expand_receive({:after, expr}, s, e) when not is_list(expr) do
-      # elixir raises here multiple_after_clauses_in_receive
-      case expr do
-        expr when not is_list(expr) ->
-          # try to recover from error by wrapping the expression in list
-          expand_receive({:after, [expr]}, s, e)
-
-        [first | discarded] ->
-          # try to recover from error by taking first clause only
-          # expand other in case there's cursor
-          {_ast, s, _e} = ElixirExpand.expand(discarded, s, e)
-          expand_receive({:after, [first]}, s, e)
-
-        [] ->
-          # try to recover from error by inserting a fake clause
-          expand_receive({:after, [{:->, [], [[0], :ok]}]}, s, e)
-      end
-    end
-
-    # with
-
-    @valid_with_opts [:do, :else]
-
-    def with(meta, args, s, e) do
-      {exprs, opts0} = ElixirUtils.split_opts(args)
-
-      # expand invalid opts in case there's cursor
-      {_ast, s, _e} = ElixirExpand.expand(opts0 |> Keyword.drop(@valid_with_opts), s, e)
-
-      opts0 = sanitize_opts(opts0, @valid_with_opts)
-      s0 = new_vars_scope(s)
-      {e_exprs, {s1, e1}} = Enum.map_reduce(exprs, {s0, e}, &expand_with/2)
-      {e_do, opts1, s2} = expand_with_do(meta, opts0, s, s1, e1)
-      {e_opts, _opts2, s3} = expand_with_else(opts1, s2, e)
-
-      {{:with, meta, e_exprs ++ [[{:do, e_do} | e_opts]]}, s3, e}
-    end
-
-    defp expand_with({:<-, meta, [left, right]}, {s, e}) do
-      {e_right, sr, er} = ElixirExpand.expand(right, s, e)
-      sm = reset_read(sr, s)
-      {[e_left], sl, el} = head([left], sm, er)
-
-      match_context_r = TypeInference.type_of(e_right, e.context)
-      vars_l_with_inferred_types = TypeInference.find_typed_vars(e_left, match_context_r, :match)
-
-      sl = State.merge_inferred_types(sl, vars_l_with_inferred_types)
-
-      {{:<-, meta, [e_left, e_right]}, {sl, el}}
-    end
-
-    defp expand_with(expr, {s, e}) do
-      {e_expr, se, ee} = ElixirExpand.expand(expr, s, e)
-      {e_expr, {se, ee}}
-    end
-
-    defp expand_with_do(_meta, opts, s, acc, e) do
-      {expr, rest_opts} = Keyword.pop(opts, :do)
-      # elixir raises here missing_option
-      # we return empty expression
-      expr = expr || []
-
-      {e_expr, s_acc, _e_acc} = ElixirExpand.expand(expr, acc, e)
-
-      {e_expr, rest_opts, remove_vars_scope(s_acc, s)}
-    end
-
-    defp expand_with_else(opts, s, e) do
-      case Keyword.pop(opts, :else) do
-        {nil, _} ->
-          {[], opts, s}
-
-        {expr, rest_opts} ->
-          pair = {:else, expr}
-
-          # no point in doing type inference here, we have no idea what data we are matching against
-          {e_pair, se} = expand_clauses(&head/3, pair, s, e)
-          {[e_pair], rest_opts, se}
-      end
-    end
-
-    # try
-
-    @valid_try_opts [:do, :rescue, :catch, :else, :after]
-
-    def try([], s, e) do
-      # elixir raises here missing_option
-      # emit a fake do block
-      try([do: []], s, e)
-    end
-
-    def try(opts, s, e) when not is_list(opts) do
-      # elixir raises here invalid_args
-      # there may be cursor
-      ElixirExpand.expand(opts, s, e)
-    end
-
-    def try(opts, s, e) do
-      # expand invalid opts in case there's cursor
-      {_ast, s, _e} = ElixirExpand.expand(opts |> Keyword.drop(@valid_try_opts), s, e)
-
-      opts = sanitize_opts(opts, @valid_try_opts)
-
-      {try_clauses, sa} =
-        Enum.map_reduce(opts, s, fn x, sa ->
-          expand_try(x, sa, e)
-        end)
-
-      {try_clauses, sa, e}
-    end
-
-    defp expand_try({:do, expr}, s, e) do
-      {e_expr, se, _ee} = ElixirExpand.expand(expr, new_vars_scope(s), e)
-      {{:do, e_expr}, remove_vars_scope(se, s)}
-    end
-
-    defp expand_try({:after, expr}, s, e) do
-      {e_expr, se, _ee} = ElixirExpand.expand(expr, new_vars_scope(s), e)
-      {{:after, e_expr}, remove_vars_scope(se, s)}
-    end
-
-    defp expand_try({:else, _} = else_clause, s, e) do
-      # TODO we could try to infer type from last try block expression
-      expand_clauses(&head/3, else_clause, s, e)
-    end
-
-    defp expand_try({:catch, _} = catch_clause, s, e) do
-      expand_clauses_with_stacktrace(&expand_catch/4, catch_clause, s, e)
-    end
-
-    defp expand_try({:rescue, _} = rescue_clause, s, e) do
-      expand_clauses_with_stacktrace(&expand_rescue/4, rescue_clause, s, e)
-    end
-
-    defp expand_clauses_with_stacktrace(fun, clauses, s, e) do
-      old_stacktrace = s.stacktrace
-      ss = %{s | stacktrace: true}
-      {ret, se} = expand_clauses(fun, clauses, ss, e)
-      {ret, %{se | stacktrace: old_stacktrace}}
-    end
-
-    defp expand_catch(meta, [{:when, when_meta, [a1, a2, a3, dh | dt]}], s, e) do
-      # elixir raises here wrong_number_of_args_for_clause
-      {_, s, _} = ElixirExpand.expand([dh | dt], s, e)
-      expand_catch(meta, [{:when, when_meta, [a1, a2, a3]}], s, e)
-    end
-
-    defp expand_catch(_meta, args = [_], s, e) do
-      # no point in doing type inference here, we have no idea what throw we caught
-      head(args, s, e)
-    end
-
-    defp expand_catch(_meta, args = [_, _], s, e) do
-      # TODO is it worth to infer type of the first arg? :error | :exit | :throw | {:EXIT, pid()}
-      head(args, s, e)
-    end
-
-    defp expand_catch(meta, [a1, a2 | d], s, e) do
-      # attempt to recover from error by taking 2 first args
-      # elixir raises here wrong_number_of_args_for_clause
-      {_, s, _} = ElixirExpand.expand(d, s, e)
-      expand_catch(meta, [a1, a2], s, e)
-    end
-
-    defp expand_rescue(_meta, [arg], s, e) do
-      # elixir is strict here and raises invalid_rescue_clause on invalid args
-      {e_arg, sa, ea} = expand_rescue(arg, s, e)
-      {[e_arg], sa, ea}
-    end
-
-    defp expand_rescue(meta, [a1 | d], s, e) do
-      # try to recover from error by taking first argument only
-      # elixir raises here wrong_number_of_args_for_clause
-      {_, s, _} = ElixirExpand.expand(d, s, e)
-      expand_rescue(meta, [a1], s, e)
-    end
-
-    # rescue var
-    defp expand_rescue({name, _, atom} = var, s, e) when is_atom(name) and is_atom(atom) do
-      {e_left, sl, el} = match(&ElixirExpand.expand/3, var, s, s, e)
-
-      match_context = {:struct, [], {:atom, Exception}, nil}
-
-      vars_with_inferred_types = TypeInference.find_typed_vars(e_left, match_context, :match)
-      sl = State.merge_inferred_types(sl, vars_with_inferred_types)
-
-      {e_left, sl, el}
-    end
-
-    # rescue Alias => _ in [Alias]
-    defp expand_rescue({:__aliases__, _, [_ | _]} = alias, s, e) do
-      expand_rescue({:in, [], [{:_, [], e.module}, alias]}, s, e)
-    end
-
-    # rescue var in _
-    defp expand_rescue(
-           {:in, _, [{name, _, var_context} = var, {:_, _, underscore_context}]},
-           s,
-           e
-         )
-         when is_atom(name) and is_atom(var_context) and is_atom(underscore_context) do
-      {e_left, sl, el} = match(&ElixirExpand.expand/3, var, s, s, e)
-
-      match_context = {:struct, [], {:atom, Exception}, nil}
-
-      vars_with_inferred_types = TypeInference.find_typed_vars(e_left, match_context, :match)
-      sl = State.merge_inferred_types(sl, vars_with_inferred_types)
-
-      {e_left, sl, el}
-    end
-
-    # rescue var in (list() or atom())
-    defp expand_rescue({:in, meta, [left, right]}, s, e) do
-      {e_left, sl, el} = match(&ElixirExpand.expand/3, left, s, s, e)
-      {e_right, sr, er} = ElixirExpand.expand(right, sl, el)
-
-      case e_left do
-        {name, _, atom} when is_atom(name) and is_atom(atom) ->
-          normalized = normalize_rescue(e_right, e)
-
-          match_context =
-            for exception <- normalized, reduce: nil do
-              nil -> {:struct, [], {:atom, exception}, nil}
-              other -> {:union, [other, {:struct, [], {:atom, exception}, nil}]}
-            end
-
-          match_context =
-            if match_context == nil do
-              {:struct, [], {:atom, Exception}, nil}
-            else
-              match_context
-            end
-
-          vars_with_inferred_types = TypeInference.find_typed_vars(e_left, match_context, :match)
-          sr = State.merge_inferred_types(sr, vars_with_inferred_types)
-
-          {{:in, meta, [e_left, normalized]}, sr, er}
-
-        _ ->
-          # elixir rejects this case, we normalize to underscore
-          {{:in, meta, [{:_, [], e.module}, normalize_rescue(e_right, e)]}, sr, er}
-      end
-    end
-
-    # rescue expr() => rescue expanded_expr()
-    defp expand_rescue({_, meta, _} = arg, s, e) do
-      # TODO how to check for cursor here?
-      case ElixirExpand.Macro.expand_once(arg, %{e | line: ElixirUtils.get_line(meta)}) do
-        ^arg ->
-          # elixir rejects this case
-          # try to recover from error by generating fake expression
-          expand_rescue({:in, meta, [arg, {:_, [], e.module}]}, s, e)
-
-        new_arg ->
-          expand_rescue(new_arg, s, e)
-      end
-    end
-
-    # rescue list() or atom() => _ in (list() or atom())
-    defp expand_rescue(arg, s, e) do
-      expand_rescue({:in, [], [{:_, [], e.module}, arg]}, s, e)
-    end
-
-    defp normalize_rescue(atom, _e) when is_atom(atom) do
-      [atom]
-    end
-
-    defp normalize_rescue(other, e) do
-      # elixir is strict here, we reject invalid nodes
-      res =
-        if is_list(other) do
-          Enum.filter(other, &is_atom/1)
-        else
-          []
-        end
-
-      if res == [] do
-        [{:_, [], e.module}]
-      else
-        res
-      end
-    end
-
-    defp expand_clauses(fun, {key, [_ | _] = clauses}, s, e) do
-      transformer = fn clause, sa ->
-        {e_clause, s_acc, _e_acc} =
-          clause(fun, clause, new_vars_scope(sa), e)
-
-        {e_clause, remove_vars_scope(s_acc, sa)}
-      end
-
-      {values, se} = Enum.map_reduce(clauses, s, transformer)
-      {{key, values}, se}
-    end
-
-    defp expand_clauses(fun, {key, expr}, s, e) do
-      # try to recover from error by wrapping the expression in a clauses list
-      # elixir raises here bad_or_missing_clauses
-      expand_clauses(fun, {key, [expr]}, s, e)
-    end
-
-    # helpers
-
-    defp sanitize_opt(opts, opt) do
-      case Keyword.fetch(opts, opt) do
-        :error -> []
-        {:ok, value} -> [{opt, value}]
-      end
-    end
-
-    defp sanitize_opts(opts, allowed) do
-      Enum.flat_map(allowed, fn opt -> sanitize_opt(opts, opt) end)
-    end
-  end
-
-  defmodule Bitstring do
-    alias ElixirSense.Core.Compiler, as: ElixirExpand
-    alias ElixirSense.Core.Compiler.Utils, as: ElixirUtils
-
-    defp expand_match(expr, {s, original_s}, e) do
-      {e_expr, se, ee} = ElixirExpand.expand(expr, s, e)
-      {e_expr, {se, original_s}, ee}
-    end
-
-    def expand(meta, args, s, e, require_size) do
-      case Map.get(e, :context) do
-        :match ->
-          {e_args, alignment, {sa, _}, ea} =
-            expand(meta, &expand_match/3, args, [], {s, s}, e, 0, require_size)
-
-          # elixir validates if there is no nested match
-
-          {{:<<>>, [{:alignment, alignment} | meta], e_args}, sa, ea}
-
-        _ ->
-          pair_s = {prepare_write(s), s}
-
-          {e_args, alignment, {sa, _}, ea} =
-            expand(meta, &ElixirExpand.expand_arg/3, args, [], pair_s, e, 0, require_size)
-
-          {{:<<>>, [{:alignment, alignment} | meta], e_args}, close_write(sa, s), ea}
-      end
-    end
-
-    def expand(_bitstr_meta, _fun, [], acc, s, e, alignment, _require_size) do
-      {Enum.reverse(acc), alignment, s, e}
-    end
-
-    def expand(
-          bitstr_meta,
-          fun,
-          [{:"::", meta, [left, right]} | t],
-          acc,
-          s,
-          e,
-          alignment,
-          require_size
-        ) do
-      {e_left, {sl, original_s}, el} = expand_expr(meta, left, fun, s, e)
-
-      match_or_require_size = require_size or is_match_size(t, el)
-      e_type = expr_type(e_left)
-
-      expect_size =
-        case e_left do
-          _ when not match_or_require_size -> :optional
-          {:^, _, [{_, _, _}]} -> {:infer, e_left}
-          _ -> :required
-        end
-
-      {e_right, e_alignment, ss, es} =
-        expand_specs(e_type, meta, right, sl, original_s, el, expect_size)
-
-      e_acc = concat_or_prepend_bitstring(meta, e_left, e_right, acc)
-
-      expand(
-        bitstr_meta,
-        fun,
-        t,
-        e_acc,
-        {ss, original_s},
-        es,
-        alignment(alignment, e_alignment),
-        require_size
-      )
-    end
-
-    def expand(bitstr_meta, fun, [h | t], acc, s, e, alignment, require_size) do
-      meta = extract_meta(h, bitstr_meta)
-      {e_left, {ss, original_s}, es} = expand_expr(meta, h, fun, s, e)
-
-      e_type = expr_type(e_left)
-      e_right = infer_spec(e_type, meta)
-
-      inferred_meta = [{:inferred_bitstring_spec, true} | meta]
-
-      e_acc =
-        concat_or_prepend_bitstring(
-          inferred_meta,
-          e_left,
-          e_right,
-          acc
-        )
-
-      expand(meta, fun, t, e_acc, {ss, original_s}, es, alignment, require_size)
-    end
-
-    defp expand_expr(
-           _meta,
-           {{:., _, [mod, :to_string]}, _, [arg]} = ast,
-           fun,
-           s,
-           %{context: context} = e
-         )
-         when context != nil and (mod == Kernel or mod == String.Chars) do
-      case fun.(arg, s, e) do
-        {ebin, se, ee} when is_binary(ebin) -> {ebin, se, ee}
-        _ -> fun.(ast, s, e)
-      end
-    end
-
-    defp expand_expr(_meta, component, fun, s, e) do
-      case fun.(component, s, e) do
-        {e_component, s, e} when is_list(e_component) or is_atom(e_component) ->
-          # elixir raises here invalid_literal
-          # try to recover from error by replacing it with ""
-          {"", s, e}
-
-        expanded ->
-          expanded
-      end
-    end
-
-    defp expand_specs(expr_type, meta, info, s, original_s, e, expect_size) do
-      default =
-        %{size: :default, unit: :default, sign: :default, type: :default, endianness: :default}
-
-      {specs, ss, es} =
-        expand_each_spec(meta, unpack_specs(info, []), default, s, original_s, e)
-
-      merged_type = type(expr_type, specs.type)
-
-      # elixir validates if unsized binary is not on the end
-
-      size_and_unit = size_and_unit(expr_type, specs.size, specs.unit)
-      alignment = compute_alignment(merged_type, specs.size, specs.unit)
-
-      maybe_inferred_size =
-        case {expect_size, merged_type, size_and_unit} do
-          {{:infer, pinned_var}, :binary, []} ->
-            [{:size, meta, [{{:., meta, [:erlang, :byte_size]}, meta, [pinned_var]}]}]
-
-          {{:infer, pinned_var}, :bitstring, []} ->
-            [{:size, meta, [{{:., meta, [:erlang, :bit_size]}, meta, [pinned_var]}]}]
-
-          _ ->
-            size_and_unit
-        end
-
-      [h | t] =
-        build_spec(
-          specs.size,
-          specs.unit,
-          merged_type,
-          specs.endianness,
-          specs.sign,
-          maybe_inferred_size
-        )
-
-      {Enum.reduce(t, h, fn i, acc -> {:-, meta, [acc, i]} end), alignment, ss, es}
-    end
-
-    defp type(:default, :default), do: :integer
-    defp type(expr_type, :default), do: expr_type
-
-    defp type(:binary, type) when type in [:binary, :bitstring, :utf8, :utf16, :utf32],
-      do: type
-
-    defp type(:bitstring, type) when type in [:binary, :bitstring], do: type
-
-    defp type(:integer, type) when type in [:integer, :float, :utf8, :utf16, :utf32],
-      do: type
-
-    defp type(:float, :float), do: :float
-    defp type(:default, type), do: type
-
-    defp type(_other, _type) do
-      # elixir raises here bittype_mismatch
-      type(:default, :default)
-    end
-
-    defp expand_each_spec(meta, [{:__cursor__, _, args} = h | t], map, s, original_s, e)
-         when is_list(args) do
-      {h, s, e} = ElixirExpand.expand(h, s, e)
-
-      args =
-        case h do
-          nil -> t
-          h -> [h | t]
-        end
-
-      expand_each_spec(meta, args, map, s, original_s, e)
-    end
-
-    defp expand_each_spec(meta, [{expr, meta_e, args} = h | t], map, s, original_s, e)
-         when is_atom(expr) do
-      case validate_spec(expr, args) do
-        {key, arg} ->
-          {value, se, ee} = expand_spec_arg(arg, s, original_s, e)
-          # elixir validates spec arg here
-          # elixir raises bittype_mismatch in some cases
-          expand_each_spec(meta, t, Map.put(map, key, value), se, original_s, ee)
-
-        :none ->
-          ha =
-            if args == nil do
-              {expr, meta_e, []}
-            else
-              h
-            end
-
-          # TODO how to check for cursor here?
-          case ElixirExpand.Macro.expand(ha, Map.put(e, :line, ElixirUtils.get_line(meta))) do
-            ^ha ->
-              # elixir raises here undefined_bittype
-              # we omit the spec
-              expand_each_spec(meta, t, map, s, original_s, e)
-
-            new_types ->
-              expand_each_spec(meta, unpack_specs(new_types, []) ++ t, map, s, original_s, e)
-          end
-      end
-    end
-
-    defp expand_each_spec(meta, [_expr | tail], map, s, original_s, e) do
-      # elixir raises undefined_bittype
-      # we skip it
-      expand_each_spec(meta, tail, map, s, original_s, e)
-    end
-
-    defp expand_each_spec(_meta, [], map, s, _original_s, e), do: {map, s, e}
-
-    defp compute_alignment(_, size, unit) when is_integer(size) and is_integer(unit),
-      do: rem(size * unit, 8)
-
-    defp compute_alignment(:default, size, unit), do: compute_alignment(:integer, size, unit)
-    defp compute_alignment(:integer, :default, unit), do: compute_alignment(:integer, 8, unit)
-    defp compute_alignment(:integer, size, :default), do: compute_alignment(:integer, size, 1)
-    defp compute_alignment(:bitstring, size, :default), do: compute_alignment(:bitstring, size, 1)
-    defp compute_alignment(:binary, size, :default), do: compute_alignment(:binary, size, 8)
-    defp compute_alignment(:binary, _, _), do: 0
-    defp compute_alignment(:float, _, _), do: 0
-    defp compute_alignment(:utf32, _, _), do: 0
-    defp compute_alignment(:utf16, _, _), do: 0
-    defp compute_alignment(:utf8, _, _), do: 0
-    defp compute_alignment(_, _, _), do: :unknown
-
-    defp alignment(left, right) when is_integer(left) and is_integer(right) do
-      rem(left + right, 8)
-    end
-
-    defp alignment(_, _), do: :unknown
-
-    defp extract_meta({_, meta, _}, _), do: meta
-    defp extract_meta(_, meta), do: meta
-
-    defp infer_spec(:bitstring, meta), do: {:bitstring, meta, nil}
-    defp infer_spec(:binary, meta), do: {:binary, meta, nil}
-    defp infer_spec(:float, meta), do: {:float, meta, nil}
-    defp infer_spec(:integer, meta), do: {:integer, meta, nil}
-    defp infer_spec(:default, meta), do: {:integer, meta, nil}
-
-    defp expr_type(integer) when is_integer(integer), do: :integer
-    defp expr_type(float) when is_float(float), do: :float
-    defp expr_type(binary) when is_binary(binary), do: :binary
-    defp expr_type({:<<>>, _, _}), do: :bitstring
-    defp expr_type(_), do: :default
-
-    defp concat_or_prepend_bitstring(_meta, {:<<>>, _, []}, _e_right, acc),
-      do: acc
-
-    defp concat_or_prepend_bitstring(
-           meta,
-           {:<<>>, parts_meta, parts} = e_left,
-           e_right,
-           acc
-         ) do
-      # elixir raises unsized_binary in some cases
-
-      case e_right do
-        {:binary, _, nil} ->
-          {alignment, alignment} = Keyword.fetch!(parts_meta, :alignment)
-
-          if is_integer(alignment) do
-            # elixir raises unaligned_binary if alignment != 0
-            Enum.reverse(parts, acc)
-          else
-            [{:"::", meta, [e_left, e_right]} | acc]
-          end
-
-        {:bitstring, _, nil} ->
-          Enum.reverse(parts, acc)
-      end
-    end
-
-    defp concat_or_prepend_bitstring(meta, e_left, e_right, acc) do
-      [{:"::", meta, [e_left, e_right]} | acc]
-    end
-
-    defp unpack_specs({:-, _, [h, t]}, acc), do: unpack_specs(h, unpack_specs(t, acc))
-
-    defp unpack_specs({:*, _, [{:_, _, atom}, unit]}, acc) when is_atom(atom),
-      do: [{:unit, [], [unit]} | acc]
-
-    defp unpack_specs({:*, _, [size, unit]}, acc),
-      do: [{:size, [], [size]}, {:unit, [], [unit]} | acc]
-
-    defp unpack_specs(size, acc) when is_integer(size), do: [{:size, [], [size]} | acc]
-
-    defp unpack_specs({expr, meta, args}, acc) when is_atom(expr) do
-      list_args =
-        cond do
-          is_atom(args) -> nil
-          is_list(args) -> args
-          true -> args
-        end
-
-      [{expr, meta, list_args} | acc]
-    end
-
-    defp unpack_specs(other, acc), do: [other | acc]
-
-    defp validate_spec(spec, []), do: validate_spec(spec, nil)
-    defp validate_spec(:big, nil), do: {:endianness, :big}
-    defp validate_spec(:little, nil), do: {:endianness, :little}
-    defp validate_spec(:native, nil), do: {:endianness, :native}
-    defp validate_spec(:size, [size]), do: {:size, size}
-    defp validate_spec(:unit, [unit]), do: {:unit, unit}
-    defp validate_spec(:integer, nil), do: {:type, :integer}
-    defp validate_spec(:float, nil), do: {:type, :float}
-    defp validate_spec(:binary, nil), do: {:type, :binary}
-    defp validate_spec(:bytes, nil), do: {:type, :binary}
-    defp validate_spec(:bitstring, nil), do: {:type, :bitstring}
-    defp validate_spec(:bits, nil), do: {:type, :bitstring}
-    defp validate_spec(:utf8, nil), do: {:type, :utf8}
-    defp validate_spec(:utf16, nil), do: {:type, :utf16}
-    defp validate_spec(:utf32, nil), do: {:type, :utf32}
-    defp validate_spec(:signed, nil), do: {:sign, :signed}
-    defp validate_spec(:unsigned, nil), do: {:sign, :unsigned}
-    defp validate_spec(_, _), do: :none
-
-    defp expand_spec_arg(expr, s, _original_s, e) when is_atom(expr) or is_integer(expr) do
-      {expr, s, e}
-    end
-
-    defp expand_spec_arg(expr, s, original_s, %{context: :match} = e) do
-      %{prematch: {pre_read, pre_counter, _} = old_pre} = s
-      %{vars: {original_read, _}} = original_s
-      new_pre = {pre_read, pre_counter, {:bitsize, original_read}}
-
-      {e_expr, se, ee} =
-        ElixirExpand.expand(expr, %{s | prematch: new_pre}, %{e | context: :guard})
-
-      {e_expr, %{se | prematch: old_pre}, %{ee | context: :match}}
-    end
-
-    defp expand_spec_arg(expr, s, original_s, e) do
-      ElixirExpand.expand(expr, reset_read(s, original_s), e)
-    end
-
-    defp size_and_unit(type, size, unit)
-         when type in [:bitstring, :binary] and (size != :default or unit != :default) do
-      # elixir raises here bittype_literal_bitstring or bittype_literal_string
-      # we don't care
-      size_and_unit(type, :default, :default)
-    end
-
-    defp size_and_unit(_expr_type, size, unit) do
-      add_arg(:unit, unit, add_arg(:size, size, []))
-    end
-
-    defp build_spec(_size, _unit, type, endianness, _sign, spec)
-         when type in [:utf8, :utf16, :utf32] do
-      # elixir raises bittype_signed if signed
-      # elixir raises bittype_utf if size specified
-      # we don't care
-
-      add_spec(type, add_spec(endianness, spec))
-    end
-
-    defp build_spec(_size, _unit, type, _endianness, _sign, spec)
-         when type in [:binary, :bitstring] do
-      # elixir raises bittype_signed if signed
-      # elixir raises bittype_mismatch if bitstring unit != 1 or default
-      # we don't care
-
-      add_spec(type, spec)
-    end
-
-    defp build_spec(size, unit, type, endianness, sign, spec)
-         when type in [:integer, :float] do
-      number_size = number_size(size, unit)
-
-      cond do
-        type == :float and is_integer(number_size) ->
-          if valid_float_size(number_size) do
-            add_spec(type, add_spec(endianness, add_spec(sign, spec)))
-          else
-            # elixir raises here bittype_float_size
-            # we fall back to 64
-            build_spec(64, :default, type, endianness, sign, spec)
-          end
-
-        size == :default and unit != :default ->
-          # elixir raises here bittype_unit
-          # we fall back to default
-          build_spec(size, :default, type, endianness, sign, spec)
-
-        true ->
-          add_spec(type, add_spec(endianness, add_spec(sign, spec)))
-      end
-    end
-
-    defp add_spec(:default, spec), do: spec
-    defp add_spec(key, spec), do: [{key, [], nil} | spec]
-
-    defp number_size(size, :default) when is_integer(size), do: size
-    defp number_size(size, unit) when is_integer(size), do: size * unit
-    defp number_size(size, _), do: size
-
-    defp valid_float_size(16), do: true
-    defp valid_float_size(32), do: true
-    defp valid_float_size(64), do: true
-    defp valid_float_size(_), do: false
-
-    defp add_arg(_key, :default, spec), do: spec
-    defp add_arg(key, arg, spec), do: [{key, [], [arg]} | spec]
-
-    defp is_match_size([_ | _], %{context: :match}), do: true
-    defp is_match_size(_, _), do: false
-  end
-
-  defmodule Fn do
-    alias ElixirSense.Core.Compiler, as: ElixirExpand
-    alias ElixirSense.Core.Compiler.Clauses, as: ElixirClauses
-    alias ElixirSense.Core.Compiler.Dispatch, as: ElixirDispatch
-    alias ElixirSense.Core.Compiler.Utils, as: ElixirUtils
-
-    def expand(meta, clauses, s, e) when is_list(clauses) do
-      transformer = fn
-        {:->, _, [_left, _right]} = clause, sa ->
-          # elixir raises defaults_in_args
-          s_reset = new_vars_scope(sa)
-
-          # no point in doing type inference here, we have no idea what the fn will be called with
-          {e_clause, s_acc, _e_acc} =
-            ElixirClauses.clause(&ElixirClauses.head/3, clause, s_reset, e)
-
-          {e_clause, remove_vars_scope(s_acc, sa)}
-      end
-
-      {e_clauses, se} = Enum.map_reduce(clauses, s, transformer)
-
-      {{:fn, meta, e_clauses}, se, e}
-    end
-
-    # Capture
-
-    def capture(meta, {:/, _, [{{:., _, [_m, f]} = dot, require_meta, []}, a]}, s, e)
-        when is_atom(f) and is_integer(a) do
-      args = args_from_arity(meta, a)
-
-      capture_require({dot, require_meta, args}, s, e, true)
-    end
-
-    def capture(meta, {:/, _, [{f, import_meta, c}, a]}, s, e)
-        when is_atom(f) and is_integer(a) and is_atom(c) do
-      args = args_from_arity(meta, a)
-      capture_import({f, import_meta, args}, s, e, true)
-    end
-
-    def capture(_meta, {{:., _, [_, fun]}, _, args} = expr, s, e)
-        when is_atom(fun) and is_list(args) do
-      capture_require(expr, s, e, is_sequential_and_not_empty(args))
-    end
-
-    def capture(meta, {{:., _, [_]}, _, args} = expr, s, e) when is_list(args) do
-      capture_expr(meta, expr, s, e, false)
-    end
-
-    def capture(meta, {:__block__, _, [expr]}, s, e) do
-      capture(meta, expr, s, e)
-    end
-
-    def capture(meta, {:__block__, _, expr}, s, e) do
-      # elixir raises block_expr_in_capture
-      # try to recover from error
-      expr =
-        case expr do
-          [] ->
-            {:"&1", meta, e.module}
-
-          list ->
-            ElixirUtils.select_with_cursor(list) || hd(list)
-        end
-
-      capture(meta, expr, s, e)
-    end
-
-    def capture(_meta, {atom, _, args} = expr, s, e) when is_atom(atom) and is_list(args) do
-      capture_import(expr, s, e, is_sequential_and_not_empty(args))
-    end
-
-    def capture(meta, {left, right}, s, e) do
-      capture(meta, {:{}, meta, [left, right]}, s, e)
-    end
-
-    def capture(meta, list, s, e) when is_list(list) do
-      capture_expr(meta, list, s, e, is_sequential_and_not_empty(list))
-    end
-
-    def capture(meta, integer, s, e) when is_integer(integer) do
-      # elixir raises here capture_arg_outside_of_capture
-      # emit fake capture
-      capture(meta, [{:&, meta, [1]}], s, e)
-    end
-
-    def capture(meta, arg, s, e) do
-      # elixir raises invalid_args_for_capture
-      # we try to transform the capture to local fun capture
-      case arg do
-        {var, _, context} when is_atom(var) and is_atom(context) ->
-          capture(meta, {:/, meta, [arg, 0]}, s, e)
-
-        _ ->
-          # try to wrap it in list
-          capture(meta, [arg], s, e)
-      end
-    end
-
-    defp capture_import({atom, import_meta, args} = expr, s, e, sequential) do
-      res =
-        if sequential do
-          ElixirDispatch.import_function(import_meta, atom, length(args), s, e)
-        else
-          false
-        end
-
-      handle_capture(res, import_meta, import_meta, expr, s, e, sequential)
-    end
-
-    defp capture_require({{:., dot_meta, [left, right]}, require_meta, args}, s, e, sequential) do
-      case escape(left, []) do
-        {esc_left, []} ->
-          {e_left, se, ee} = ElixirExpand.expand(esc_left, s, e)
-
-          res =
-            if sequential do
-              case e_left do
-                {name, _, context} when is_atom(name) and is_atom(context) ->
-                  {:remote, e_left, right, length(args)}
-
-                _ when is_atom(e_left) ->
-                  ElixirDispatch.require_function(
-                    require_meta,
-                    e_left,
-                    right,
-                    length(args),
-                    s,
-                    ee
-                  )
-
-                _ ->
-                  false
-              end
-            else
-              false
-            end
-
-          dot = {{:., dot_meta, [e_left, right]}, require_meta, args}
-          handle_capture(res, require_meta, dot_meta, dot, se, ee, sequential)
-
-        {esc_left, escaped} ->
-          dot = {{:., dot_meta, [esc_left, right]}, require_meta, args}
-          capture_expr(require_meta, dot, s, e, escaped, sequential)
-      end
-    end
-
-    defp handle_capture(false, meta, _dot_meta, expr, s, e, sequential) do
-      capture_expr(meta, expr, s, e, sequential)
-    end
-
-    defp handle_capture(local_or_remote, meta, dot_meta, _expr, s, e, _sequential) do
-      {local_or_remote, meta, dot_meta, s, e}
-    end
-
-    defp capture_expr(meta, expr, s, e, sequential) do
-      capture_expr(meta, expr, s, e, [], sequential)
-    end
-
-    defp capture_expr(meta, expr, s, e, escaped, sequential) do
-      case escape(expr, escaped) do
-        {e_expr, []} when not sequential ->
-          # elixir raises here invalid_args_for_capture
-          # we emit fn without args
-          fn_expr = {:fn, meta, [{:->, meta, [[], e_expr]}]}
-          {:expand, fn_expr, s, e}
-
-        {e_expr, e_dict} ->
-          # elixir raises capture_arg_without_predecessor here
-          # if argument vars are not consecutive
-          e_vars = Enum.map(e_dict, &elem(&1, 1))
-          fn_expr = {:fn, meta, [{:->, meta, [e_vars, e_expr]}]}
-          {:expand, fn_expr, s, e}
-      end
-    end
-
-    defp escape({:&, meta, [pos]}, dict) when is_integer(pos) and pos > 0 do
-      # This might pollute user space but is unlikely because variables
-      # named :"&1" are not valid syntax.
-      var = {:"&#{pos}", meta, nil}
-      {var, :orddict.store(pos, var, dict)}
-
-      case :orddict.find(pos, dict) do
-        {:ok, var} ->
-          {var, dict}
-
-        :error ->
-          # elixir uses here elixir_module:next_counter(?key(E, module))
-          # but we are not compiling and do not need to keep count in module scope
-          # elixir 1.17 also renames the var to `capture`
-          next = System.unique_integer()
-          var = {:"&#{pos}", [{:counter, next} | meta], nil}
-          {var, :orddict.store(pos, var, dict)}
-      end
-    end
-
-    defp escape({:&, meta, [pos]}, dict) when is_integer(pos) do
-      # elixir raises here invalid_arity_for_capture
-      # we substitute arg number
-      escape({:&, meta, [1]}, dict)
-    end
-
-    defp escape({:&, _meta, args}, dict) do
-      # elixir raises here nested_capture
-      # try to recover from error by dropping &
-      escape(args, dict)
-    end
-
-    defp escape({left, meta, right}, dict0) do
-      {t_left, dict1} = escape(left, dict0)
-      {t_right, dict2} = escape(right, dict1)
-      {{t_left, meta, t_right}, dict2}
-    end
-
-    defp escape({left, right}, dict0) do
-      {t_left, dict1} = escape(left, dict0)
-      {t_right, dict2} = escape(right, dict1)
-      {{t_left, t_right}, dict2}
-    end
-
-    defp escape(list, dict) when is_list(list) do
-      Enum.map_reduce(list, dict, fn x, acc -> escape(x, acc) end)
-    end
-
-    defp escape(other, dict) do
-      {other, dict}
-    end
-
-    defp args_from_arity(_meta, 0), do: []
-
-    defp args_from_arity(meta, a) when is_integer(a) and a >= 1 and a <= 255 do
-      for x <- 1..a do
-        {:&, meta, [x]}
-      end
-    end
-
-    defp args_from_arity(_meta, _a) do
-      # elixir raises invalid_arity_for_capture
-      []
-    end
-
-    defp is_sequential_and_not_empty([]), do: false
-    defp is_sequential_and_not_empty(list), do: is_sequential(list, 1)
-
-    defp is_sequential([{:&, _, [int]} | t], int), do: is_sequential(t, int + 1)
-    defp is_sequential([], _int), do: true
-    defp is_sequential(_, _int), do: false
-  end
-
-  defmodule Quote do
-    alias ElixirSense.Core.Compiler.Dispatch, as: ElixirDispatch
-
-    defstruct line: false,
-              file: nil,
-              context: nil,
-              op: :none,
-              aliases_hygiene: nil,
-              imports_hygiene: nil,
-              unquote: true,
-              generated: false
-
-    def fun_to_quoted(function) do
-      {:module, module} = :erlang.fun_info(function, :module)
-      {:name, name} = :erlang.fun_info(function, :name)
-      {:arity, arity} = :erlang.fun_info(function, :arity)
-
-      {:&, [], [{:/, [], [{{:., [], [module, name]}, [{:no_parens, true}], []}, arity]}]}
-    end
-
-    def has_unquotes(ast), do: has_unquotes(ast, 0)
-
-    def has_unquotes({:quote, _, [child]}, quote_level) do
-      has_unquotes(child, quote_level + 1)
-    end
-
-    def has_unquotes({:quote, _, [quote_opts, child]}, quote_level) do
-      case disables_unquote(quote_opts) do
-        true -> false
-        _ -> has_unquotes(child, quote_level + 1)
-      end
-    end
-
-    def has_unquotes({unquote, _, [child]}, quote_level)
-        when unquote in [:unquote, :unquote_splicing] do
-      case quote_level do
-        0 -> true
-        _ -> has_unquotes(child, quote_level - 1)
-      end
-    end
-
-    def has_unquotes({{:., _, [_, :unquote]}, _, [_]}, _), do: true
-    def has_unquotes({var, _, ctx}, _) when is_atom(var) and is_atom(ctx), do: false
-
-    def has_unquotes({name, _, args}, quote_level) when is_list(args) do
-      has_unquotes(name) or Enum.any?(args, fn child -> has_unquotes(child, quote_level) end)
-    end
-
-    def has_unquotes({left, right}, quote_level) do
-      has_unquotes(left, quote_level) or has_unquotes(right, quote_level)
-    end
-
-    def has_unquotes(list, quote_level) when is_list(list) do
-      Enum.any?(list, fn child -> has_unquotes(child, quote_level) end)
-    end
-
-    def has_unquotes(_other, _), do: false
-
-    defp disables_unquote([{:unquote, false} | _]), do: true
-    defp disables_unquote([{:bind_quoted, _} | _]), do: true
-    defp disables_unquote([_h | t]), do: disables_unquote(t)
-    defp disables_unquote(_), do: false
-
-    def build(meta, line, file, context, unquote, generated, e) do
-      acc0 = []
-
-      {v_line, acc1} = validate_compile(meta, :line, line, acc0)
-      {v_file, acc2} = validate_compile(meta, :file, file, acc1)
-      {v_context, acc3} = validate_compile(meta, :context, context, acc2)
-
-      unquote = validate_runtime(:unquote, unquote)
-      generated = validate_runtime(:generated, generated)
-
-      q = %__MODULE__{
-        op: :add_context,
-        aliases_hygiene: e,
-        imports_hygiene: e,
-        line: v_line,
-        file: v_file,
-        unquote: unquote,
-        context: v_context,
-        generated: generated
-      }
-
-      {q, v_context, acc3}
-    end
-
-    defp validate_compile(_meta, :line, value, acc) when is_boolean(value) do
-      {value, acc}
-    end
-
-    defp validate_compile(_meta, :file, nil, acc) do
-      {nil, acc}
-    end
-
-    defp validate_compile(meta, key, value, acc) do
-      case is_valid(key, value) do
-        true ->
-          {value, acc}
-
-        false ->
-          var = {key, meta, __MODULE__}
-          call = {{:., meta, [__MODULE__, :validate_runtime]}, meta, [key, value]}
-          {var, [{:=, meta, [var, call]} | acc]}
-      end
-    end
-
-    defp validate_runtime(key, value) do
-      case is_valid(key, value) do
-        true ->
-          value
-
-        false ->
-          # elixir raises here invalid runtime value for option
-          default(key)
-      end
-    end
-
-    defp is_valid(:line, line), do: is_integer(line)
-    defp is_valid(:file, file), do: is_binary(file)
-    defp is_valid(:context, context), do: is_atom(context) and context != nil
-    defp is_valid(:generated, generated), do: is_boolean(generated)
-    defp is_valid(:unquote, unquote), do: is_boolean(unquote)
-    defp default(:unquote), do: true
-    defp default(:generated), do: false
-
-    def escape(expr, op, unquote) do
-      do_quote(
-        expr,
-        %__MODULE__{
-          line: true,
-          file: nil,
-          op: op,
-          unquote: unquote
-        }
-      )
-    end
-
-    def quote({:unquote_splicing, _, [_]} = expr, %__MODULE__{unquote: true} = q) do
-      # elixir raises here unquote_splicing only works inside arguments and block contexts
-      # try to recover from error by wrapping it in block
-      __MODULE__.quote({:__block__, [], [expr]}, q)
-    end
-
-    def quote(expr, q) do
-      do_quote(expr, q)
-    end
-
-    # quote/unquote
-
-    defp do_quote({:quote, meta, [arg]}, q) when is_list(meta) do
-      t_arg = do_quote(arg, %__MODULE__{q | unquote: false})
-
-      new_meta =
-        case q do
-          %__MODULE__{op: :add_context, context: context} ->
-            keystore(:context, meta, context)
-
-          _ ->
-            meta
-        end
-
-      {:{}, [], [:quote, meta(new_meta, q), [t_arg]]}
-    end
-
-    defp do_quote({:quote, meta, [opts, arg]}, q) when is_list(meta) do
-      t_opts = do_quote(opts, q)
-      t_arg = do_quote(arg, %__MODULE__{q | unquote: false})
-
-      new_meta =
-        case q do
-          %__MODULE__{op: :add_context, context: context} ->
-            keystore(:context, meta, context)
-
-          _ ->
-            meta
-        end
-
-      {:{}, [], [:quote, meta(new_meta, q), [t_opts, t_arg]]}
-    end
-
-    defp do_quote({:unquote, meta, [expr]}, %__MODULE__{unquote: true}) when is_list(meta),
-      do: expr
-
-    # Aliases
-
-    defp do_quote({:__aliases__, meta, [h | t] = list}, %__MODULE__{aliases_hygiene: e = %{}} = q)
-         when is_atom(h) and h != Elixir and is_list(meta) do
-      annotation =
-        case NormalizedMacroEnv.expand_alias(e, meta, list, trace: false) do
-          {:alias, atom} -> atom
-          :error -> false
-        end
-
-      alias_meta = keystore(:alias, Keyword.delete(meta, :counter), annotation)
-      do_quote_tuple(:__aliases__, alias_meta, [h | t], q)
-    end
-
-    # Vars
-
-    defp do_quote({name, meta, nil}, %__MODULE__{op: :add_context} = q)
-         when is_atom(name) and is_list(meta) do
-      import_meta =
-        case q.imports_hygiene do
-          nil -> meta
-          e -> import_meta(meta, name, 0, q, e)
-        end
-
-      {:{}, [], [name, meta(import_meta, q), q.context]}
-    end
-
-    # cursor
-
-    defp do_quote(
-           {:__cursor__, meta, args},
-           %__MODULE__{unquote: _}
-         )
-         when is_list(args) do
-      # emit cursor as is regardless of unquote
-      {:__cursor__, meta, args}
-    end
-
-    # Unquote
-
-    defp do_quote(
-           {{{:., meta, [left, :unquote]}, _, [expr]}, _, args},
-           %__MODULE__{unquote: true} = q
-         )
-         when is_list(meta) do
-      do_quote_call(left, meta, expr, args, q)
-    end
-
-    defp do_quote({{:., meta, [left, :unquote]}, _, [expr]}, %__MODULE__{unquote: true} = q)
-         when is_list(meta) do
-      do_quote_call(left, meta, expr, nil, q)
-    end
-
-    # Imports
-
-    defp do_quote(
-           {:&, meta, [{:/, _, [{f, _, c}, a]}] = args},
-           %__MODULE__{imports_hygiene: e = %{}} = q
-         )
-         when is_atom(f) and is_integer(a) and is_atom(c) and is_list(meta) do
-      new_meta =
-        case ElixirDispatch.find_import(meta, f, a, e) do
-          false ->
-            meta
-
-          receiver ->
-            keystore(:context, keystore(:imports, meta, [{a, receiver}]), q.context)
-        end
-
-      do_quote_tuple(:&, new_meta, args, q)
-    end
-
-    defp do_quote({name, meta, args_or_context}, %__MODULE__{imports_hygiene: e = %{}} = q)
-         when is_atom(name) and is_list(meta) and
-                (is_list(args_or_context) or is_atom(args_or_context)) do
-      arity =
-        case args_or_context do
-          args when is_list(args) -> length(args)
-          context when is_atom(context) -> 0
-        end
-
-      import_meta = import_meta(meta, name, arity, q, e)
-      annotated = annotate({name, import_meta, args_or_context}, q.context)
-      do_quote_tuple(annotated, q)
-    end
-
-    # Two-element tuples
-
-    defp do_quote({left, right}, %__MODULE__{unquote: true} = q)
-         when is_tuple(left) and elem(left, 0) == :unquote_splicing and
-                is_tuple(right) and elem(right, 0) == :unquote_splicing do
-      do_quote({:{}, [], [left, right]}, q)
-    end
-
-    defp do_quote({left, right}, q) do
-      t_left = do_quote(left, q)
-      t_right = do_quote(right, q)
-      {t_left, t_right}
-    end
-
-    # Everything else
-
-    defp do_quote(other, q = %{op: op}) when op != :add_context do
-      do_escape(other, q)
-    end
-
-    defp do_quote({_, _, _} = tuple, q) do
-      annotated = annotate(tuple, q.context)
-      do_quote_tuple(annotated, q)
-    end
-
-    defp do_quote([], _), do: []
-
-    defp do_quote([h | t], %__MODULE__{unquote: false} = q) do
-      head_quoted = do_quote(h, q)
-      do_quote_simple_list(t, head_quoted, q)
-    end
-
-    defp do_quote([h | t], q) do
-      do_quote_tail(:lists.reverse(t, [h]), q)
-    end
-
-    defp do_quote(other, _), do: other
-
-    defp import_meta(meta, name, arity, q, e) do
-      case Keyword.get(meta, :imports, false) == false &&
-             ElixirDispatch.find_imports(meta, name, e) do
-        [_ | _] = imports ->
-          keystore(:imports, keystore(:context, meta, q.context), imports)
-
-        _ ->
-          case arity == 1 && Keyword.fetch(meta, :ambiguous_op) do
-            {:ok, nil} ->
-              keystore(:ambiguous_op, meta, q.context)
-
-            _ ->
-              meta
-          end
-      end
-    end
-
-    defp do_quote_call(left, meta, expr, args, q) do
-      all = [left, {:unquote, meta, [expr]}, args, q.context]
-      tall = Enum.map(all, fn x -> do_quote(x, q) end)
-      {{:., meta, [:elixir_quote, :dot]}, meta, [meta(meta, q) | tall]}
-    end
-
-    defp do_quote_tuple({left, meta, right}, q) do
-      do_quote_tuple(left, meta, right, q)
-    end
-
-    defp do_quote_tuple(left, meta, right, q) do
-      t_left = do_quote(left, q)
-      t_right = do_quote(right, q)
-      {:{}, [], [t_left, meta(meta, q), t_right]}
-    end
-
-    defp do_quote_simple_list([], prev, _), do: [prev]
-
-    defp do_quote_simple_list([h | t], prev, q) do
-      [prev | do_quote_simple_list(t, do_quote(h, q), q)]
-    end
-
-    defp do_quote_simple_list(other, prev, q) do
-      [{:|, [], [prev, do_quote(other, q)]}]
-    end
-
-    defp do_quote_tail(
-           [{:|, meta, [{:unquote_splicing, _, [left]}, right]} | t],
-           %__MODULE__{unquote: true} = q
-         ) do
-      tt = do_quote_splice(t, q, [], [])
-      tr = do_quote(right, q)
-      do_runtime_list(meta, :tail_list, [left, tr, tt])
-    end
-
-    defp do_quote_tail(list, q) do
-      do_quote_splice(list, q, [], [])
-    end
-
-    defp do_quote_splice(
-           [{:unquote_splicing, meta, [expr]} | t],
-           %__MODULE__{unquote: true} = q,
-           buffer,
-           acc
-         ) do
-      runtime = do_runtime_list(meta, :list, [expr, do_list_concat(buffer, acc)])
-      do_quote_splice(t, q, [], runtime)
-    end
-
-    defp do_quote_splice([h | t], q, buffer, acc) do
-      th = do_quote(h, q)
-      do_quote_splice(t, q, [th | buffer], acc)
-    end
-
-    defp do_quote_splice([], _q, buffer, acc) do
-      do_list_concat(buffer, acc)
-    end
-
-    defp do_list_concat(left, []), do: left
-    defp do_list_concat([], right), do: right
-
-    defp do_list_concat(left, right) do
-      {{:., [], [:erlang, :++]}, [], [left, right]}
-    end
-
-    defp do_runtime_list(meta, fun, args) do
-      {{:., meta, [:elixir_quote, fun]}, meta, args}
-    end
-
-    defp meta(meta, q) do
-      generated(keep(Keyword.delete(meta, :column), q), q)
-    end
-
-    defp generated(meta, %__MODULE__{generated: true}), do: [{:generated, true} | meta]
-    defp generated(meta, %__MODULE__{generated: false}), do: meta
-
-    defp keep(meta, %__MODULE__{file: nil, line: line}) do
-      line(meta, line)
-    end
-
-    defp keep(meta, %__MODULE__{file: file, line: true}) do
-      case Keyword.pop(meta, :line) do
-        {nil, _} ->
-          [{:keep, {file, 0}} | meta]
-
-        {line, meta_no_line} ->
-          [{:keep, {file, line}} | meta_no_line]
-      end
-    end
-
-    defp keep(meta, %__MODULE__{file: file, line: false}) do
-      [{:keep, {file, 0}} | Keyword.delete(meta, :line)]
-    end
-
-    defp keep(meta, %__MODULE__{file: file, line: line}) do
-      [{:keep, {file, line}} | Keyword.delete(meta, :line)]
-    end
-
-    defp line(meta, true), do: meta
-
-    defp line(meta, false) do
-      Keyword.delete(meta, :line)
-    end
-
-    defp line(meta, line) do
-      keystore(:line, meta, line)
-    end
-
-    defguardp defs(kind) when kind in [:def, :defp, :defmacro, :defmacrop, :@]
-    defguardp lexical(kind) when kind in [:import, :alias, :require]
-
-    defp annotate({def, meta, [h | t]}, context) when defs(def) do
-      {def, meta, [annotate_def(h, context) | t]}
-    end
-
-    defp annotate({{:., _, [_, def]} = target, meta, [h | t]}, context) when defs(def) do
-      {target, meta, [annotate_def(h, context) | t]}
-    end
-
-    defp annotate({lexical, meta, [_ | _] = args}, context) when lexical(lexical) do
-      new_meta = keystore(:context, Keyword.delete(meta, :counter), context)
-      {lexical, new_meta, args}
-    end
-
-    defp annotate(tree, _context), do: tree
-
-    defp annotate_def({:when, meta, [left, right]}, context) do
-      {:when, meta, [annotate_def(left, context), right]}
-    end
-
-    defp annotate_def({fun, meta, args}, context) do
-      {fun, keystore(:context, meta, context), args}
-    end
-
-    defp annotate_def(other, _context), do: other
-
-    defp do_escape({left, meta, right}, q = %{op: :prune_metadata}) when is_list(meta) do
-      tm = for {k, v} <- meta, k == :no_parens or k == :line, do: {k, v}
-      tl = do_quote(left, q)
-      tr = do_quote(right, q)
-      {:{}, [], [tl, tm, tr]}
-    end
-
-    defp do_escape(tuple, q) when is_tuple(tuple) do
-      tt = do_quote(Tuple.to_list(tuple), q)
-      {:{}, [], tt}
-    end
-
-    defp do_escape(bitstring, _) when is_bitstring(bitstring) do
-      case Bitwise.band(bit_size(bitstring), 7) do
-        0 ->
-          bitstring
-
-        size ->
-          <<bits::size(size), bytes::binary>> = bitstring
-
-          {:<<>>, [],
-           [{:"::", [], [bits, {size, [], [size]}]}, {:"::", [], [bytes, {:binary, [], nil}]}]}
-      end
-    end
-
-    defp do_escape(map, q) when is_map(map) do
-      tt = do_quote(Enum.sort(Map.to_list(map)), q)
-      {:%{}, [], tt}
-    end
-
-    defp do_escape([], _), do: []
-
-    defp do_escape([h | t], %__MODULE__{unquote: false} = q) do
-      do_quote_simple_list(t, do_quote(h, q), q)
-    end
-
-    defp do_escape([h | t], q) do
-      # The improper case is inefficient, but improper lists are rare.
-      try do
-        l = Enum.reverse(t, [h])
-        do_quote_tail(l, q)
-      catch
-        _ ->
-          {l, r} = reverse_improper(t, [h])
-          tl = do_quote_splice(l, q, [], [])
-          tr = do_quote(r, q)
-          update_last(tl, fn x -> {:|, [], [x, tr]} end)
-      end
-    end
-
-    defp do_escape(other, _) when is_number(other) or is_pid(other) or is_atom(other),
-      do: other
-
-    defp do_escape(fun, _) when is_function(fun) do
-      case {Function.info(fun, :env), Function.info(fun, :type)} do
-        {{:env, []}, {:type, :external}} ->
-          fun_to_quoted(fun)
-
-        _ ->
-          # elixir raises here ArgumentError
-          nil
-      end
-    end
-
-    defp do_escape(_other, _) do
-      # elixir raises here ArgumentError
-      nil
-    end
-
-    defp reverse_improper([h | t], acc), do: reverse_improper(t, [h | acc])
-    defp reverse_improper([], acc), do: acc
-    defp reverse_improper(t, acc), do: {acc, t}
-    defp update_last([], _), do: []
-    defp update_last([h], f), do: [f.(h)]
-    defp update_last([h | t], f), do: [h | update_last(t, f)]
-
-    defp keystore(_key, meta, value) when value == nil do
-      meta
-    end
-
-    defp keystore(key, meta, value) do
-      :lists.keystore(key, 1, meta, {key, value})
-    end
-  end
-
-  defmodule Dispatch do
-    alias ElixirSense.Core.Compiler.Rewrite, as: ElixirRewrite
-    import :ordsets, only: [is_element: 2]
-
-    def find_import(meta, name, arity, e) do
-      tuple = {name, arity}
-
-      case find_import_by_name_arity(meta, tuple, [], e) do
-        {:function, receiver} ->
-          # TODO trace call?
-          # TODO address when https://github.com/elixir-lang/elixir/issues/13878 is resolved
-          # ElixirEnv.trace({:imported_function, meta, receiver, name, arity}, e)
-          receiver
-
-        {:macro, receiver} ->
-          # TODO trace call?
-          # ElixirEnv.trace({:imported_macro, meta, receiver, name, arity}, e)
-          receiver
-
-        {:ambiguous, [head | _]} ->
-          # elixir raises here, we choose first one
-          # TODO trace call?
-          head
-
-        _ ->
-          false
-      end
-    end
-
-    def find_imports(meta, name, e) do
-      funs = e.functions
-      macs = e.macros
-
-      acc0 = %{}
-      acc1 = find_imports_by_name(funs, acc0, name, meta, e)
-      acc2 = find_imports_by_name(macs, acc1, name, meta, e)
-
-      imports = acc2 |> Map.to_list() |> Enum.sort()
-      # trace_import_quoted(imports, meta, name, e)
-      imports
-    end
-
-    def import_function(meta, name, arity, s, e) do
-      tuple = {name, arity}
-
-      case find_import_by_name_arity(meta, tuple, [], e) do
-        {:function, receiver} ->
-          remote_function(meta, receiver, name, arity, e)
-
-        {:macro, _receiver} ->
-          false
-
-        {:import, receiver} ->
-          require_function(meta, receiver, name, arity, s, e)
-
-        {:ambiguous, [first | _]} ->
-          # elixir raises here, we return first matching
-          require_function(meta, first, name, arity, s, e)
-
-        false ->
-          if Macro.special_form?(name, arity) do
-            false
-          else
-            function = e.function
-
-            mfa = {e.module, name, arity}
-
-            if function != nil and function != tuple and
-                 Enum.any?(s.mods_funs_to_positions, fn {key, info} ->
-                   key == mfa and State.ModFunInfo.get_category(info) == :macro
-                 end) do
-              false
-            else
-              {:local, name, arity}
-            end
-          end
-      end
-    end
-
-    def require_function(meta, receiver, name, arity, s, e) do
-      required = receiver in e.requires
-
-      if is_macro(name, arity, receiver, required, s) do
-        false
-      else
-        remote_function(meta, receiver, name, arity, e)
-      end
-    end
-
-    defp remote_function(_meta, receiver, name, arity, _e) do
-      case ElixirRewrite.inline(receiver, name, arity) do
-        {ar, an} -> {:remote, ar, an, arity}
-        false -> {:remote, receiver, name, arity}
-      end
-    end
-
-    def find_imports_by_name([{mod, imports} | mod_imports], acc, name, meta, e) do
-      new_acc = find_imports_by_name(name, imports, acc, mod, meta, e)
-      find_imports_by_name(mod_imports, new_acc, name, meta, e)
-    end
-
-    def find_imports_by_name([], acc, _name, _meta, _e), do: acc
-
-    def find_imports_by_name(name, [{name, arity} | imports], acc, mod, meta, e) do
-      case Map.get(acc, arity) do
-        nil ->
-          find_imports_by_name(name, imports, Map.put(acc, arity, mod), mod, meta, e)
-
-        _other_mod ->
-          # elixir raises here ambiguous_call
-          find_imports_by_name(name, imports, acc, mod, meta, e)
-      end
-    end
-
-    def find_imports_by_name(name, [{import_name, _} | imports], acc, mod, meta, e)
-        when name > import_name do
-      find_imports_by_name(name, imports, acc, mod, meta, e)
-    end
-
-    def find_imports_by_name(_name, _imports, acc, _mod, _meta, _e), do: acc
-
-    defp find_import_by_name_arity(meta, {_name, arity} = tuple, extra, e) do
-      case is_import(meta, arity) do
-        {:import, _} = import_res ->
-          import_res
-
-        false ->
-          funs = e.functions
-          macs = extra ++ e.macros
-          fun_match = find_import_by_name_arity(tuple, funs)
-          mac_match = find_import_by_name_arity(tuple, macs)
-
-          case {fun_match, mac_match} do
-            {[], [receiver]} ->
-              {:macro, receiver}
-
-            {[receiver], []} ->
-              {:function, receiver}
-
-            {[], []} ->
-              false
-
-            _ ->
-              {:ambiguous, fun_match ++ mac_match}
-          end
-      end
-    end
-
-    defp find_import_by_name_arity(tuple, list) do
-      for {receiver, set} <- list, is_element(tuple, set), do: receiver
-    end
-
-    defp is_import(meta, arity) do
-      with {:ok, imports = [_ | _]} <- Keyword.fetch(meta, :imports),
-           {:ok, _} <- Keyword.fetch(meta, :context),
-           {_arity, receiver} <- :lists.keyfind(arity, 1, imports) do
-        {:import, receiver}
-      else
-        _ -> false
-      end
-    end
-
-    defp is_macro(_name, _arity, _module, false, _s), do: false
-
-    defp is_macro(name, arity, receiver, true, s) do
-      mfa = {receiver, name, arity}
-
-      Enum.any?(s.mods_funs_to_positions, fn {key, info} ->
-        key == mfa and State.ModFunInfo.get_category(info) == :macro
-      end) ||
-        try do
-          macros = receiver.__info__(:macros)
-          {name, arity} in macros
-        rescue
-          _error -> false
-        end
-    end
-  end
-
-  defmodule Map do
-    alias ElixirSense.Core.Compiler, as: ElixirExpand
-
-    def expand_struct(meta, left, {:%{}, map_meta, map_args}, s, %{context: context} = e) do
-      clean_map_args = clean_struct_key_from_map_args(map_args)
-
-      {[e_left, e_right], se, ee} =
-        ElixirExpand.expand_args([left, {:%{}, map_meta, clean_map_args}], s, e)
-
-      case validate_struct(e_left, context) do
-        true when is_atom(e_left) ->
-          # TODO register alias/struct
-          case extract_struct_assocs(e_right) do
-            {:expand, map_meta, assocs} when context != :match ->
-              assoc_keys = Enum.map(assocs, fn {k, _} -> k end)
-              struct = load_struct(e_left, [assocs], se, ee)
-              keys = [:__struct__ | assoc_keys]
-              without_keys = Elixir.Map.drop(struct, keys)
-
-              struct_assocs =
-                ElixirExpand.Macro.escape(Enum.sort(Elixir.Map.to_list(without_keys)))
-
-              {{:%, meta, [e_left, {:%{}, map_meta, struct_assocs ++ assocs}]}, se, ee}
-
-            {_, _, _assocs} ->
-              # elixir validates assocs against struct keys
-              # we don't need to validate keys
-              {{:%, meta, [e_left, e_right]}, se, ee}
-          end
-
-        _ ->
-          # elixir raises invalid_struct_name if validate_struct returns false
-          {{:%, meta, [e_left, e_right]}, se, ee}
-      end
-    end
-
-    def expand_struct(meta, left, right, s, e) do
-      # elixir raises here non_map_after_struct
-      # try to recover from error by wrapping the expression in map
-      expand_struct(meta, left, wrap_in_fake_map(right), s, e)
-    end
-
-    defp wrap_in_fake_map(right) do
-      map_args =
-        case right do
-          list when is_list(list) ->
-            if Keyword.keyword?(list) do
-              list
-            else
-              [__fake_key__: list]
-            end
-
-          _ ->
-            [__fake_key__: right]
-        end
-
-      {:%{}, [], map_args}
-    end
-
-    def expand_map(meta, [{:|, update_meta, [left, right]}], s, e) do
-      # elixir raises update_syntax_in_wrong_context if e.context is not nil
-      {[e_left | e_right], se, ee} = ElixirExpand.expand_args([left | right], s, e)
-      e_right = sanitize_kv(e_right, e)
-      {{:%{}, meta, [{:|, update_meta, [e_left, e_right]}]}, se, ee}
-    end
-
-    def expand_map(meta, args, s, e) do
-      {e_args, se, ee} = ElixirExpand.expand_args(args, s, e)
-      e_args = sanitize_kv(e_args, e)
-      {{:%{}, meta, e_args}, se, ee}
-    end
-
-    defp clean_struct_key_from_map_args([{:|, pipe_meta, [left, map_assocs]}]) do
-      [{:|, pipe_meta, [left, delete_struct_key(map_assocs)]}]
-    end
-
-    defp clean_struct_key_from_map_args(map_assocs) do
-      delete_struct_key(map_assocs)
-    end
-
-    defp sanitize_kv(kv, %{context: context}) do
-      Enum.filter(kv, fn
-        {k, _v} ->
-          if context == :match do
-            validate_match_key(k)
-          else
-            true
-          end
-
-        _ ->
-          false
-      end)
-    end
-
-    defp validate_match_key({name, _, context})
-         when is_atom(name) and is_atom(context) do
-      # elixir raises here invalid_variable_in_map_key_match
-      false
-    end
-
-    defp validate_match_key({:"::", _, [left, _]}) do
-      validate_match_key(left)
-    end
-
-    defp validate_match_key({:^, _, [{name, _, context}]})
-         when is_atom(name) and is_atom(context),
-         do: true
-
-    defp validate_match_key({:%{}, _, [_ | _]}), do: true
-
-    defp validate_match_key({left, _, right}) do
-      validate_match_key(left) and validate_match_key(right)
-    end
-
-    defp validate_match_key({left, right}) do
-      validate_match_key(left) and validate_match_key(right)
-    end
-
-    defp validate_match_key(list) when is_list(list) do
-      Enum.all?(list, &validate_match_key/1)
-    end
-
-    defp validate_match_key(_), do: true
-
-    defp validate_struct({:^, _, [{var, _, ctx}]}, :match) when is_atom(var) and is_atom(ctx),
-      do: true
-
-    defp validate_struct({var, _meta, ctx}, :match) when is_atom(var) and is_atom(ctx), do: true
-    defp validate_struct(atom, _) when is_atom(atom), do: true
-    defp validate_struct(_, _), do: false
-
-    defp sanitize_assocs(list) do
-      Enum.filter(list, &match?({k, _} when is_atom(k), &1))
-    end
-
-    defp extract_struct_assocs({:%{}, meta, [{:|, _, [_, assocs]}]}) do
-      {:update, meta, delete_struct_key(sanitize_assocs(assocs))}
-    end
-
-    defp extract_struct_assocs({:%{}, meta, assocs}) do
-      {:expand, meta, delete_struct_key(sanitize_assocs(assocs))}
-    end
-
-    defp extract_struct_assocs(right) do
-      # elixir raises here non_map_after_struct
-      # try to recover from error by wrapping the expression in map
-      extract_struct_assocs(wrap_in_fake_map(right))
-    end
-
-    defp delete_struct_key(assocs) do
-      Keyword.delete(assocs, :__struct__)
-    end
-
-    def load_struct(name, assocs, s, _e) do
-      case s.structs[name] do
-        nil ->
-          try do
-            apply(name, :__struct__, assocs)
-          else
-            %{:__struct__ => ^name} = struct ->
-              struct
-
-            _ ->
-              # recover from invalid return value
-              [__struct__: name] |> merge_assocs(assocs)
-          rescue
-            _ ->
-              # recover from error by building the fake struct
-              [__struct__: name] |> merge_assocs(assocs)
-          end
-
-        info ->
-          info.fields |> merge_assocs(assocs)
-      end
-    end
-
-    defp merge_assocs(fields, []) do
-      fields |> Elixir.Map.new()
-    end
-
-    defp merge_assocs(fields, [assocs]) do
-      fields |> Keyword.merge(assocs) |> Elixir.Map.new()
-    end
-  end
-
-  defmodule Rewrite do
-    def inline(module, fun, arity) do
-      :elixir_rewrite.inline(module, fun, arity)
-    end
-
-    def rewrite(context, receiver, dot_meta, right, meta, e_args, s) do
-      do_rewrite(context, receiver, dot_meta, right, meta, e_args, s)
-    end
-
-    defp do_rewrite(_, :erlang, _, :+, _, [arg], _s) when is_number(arg), do: {:ok, arg}
-
-    defp do_rewrite(_, :erlang, _, :-, _, [arg], _s) when is_number(arg), do: {:ok, -arg}
-
-    defp do_rewrite(:match, receiver, dot_meta, right, meta, e_args, _s) do
-      :elixir_rewrite.match_rewrite(receiver, dot_meta, right, meta, e_args)
-    end
-
-    if Version.match?(System.version(), "< 1.14.0") do
-      defp do_rewrite(:guard, receiver, dot_meta, right, meta, e_args, s) do
-        :elixir_rewrite.guard_rewrite(receiver, dot_meta, right, meta, e_args)
-      end
-    else
-      defp do_rewrite(:guard, receiver, dot_meta, right, meta, e_args, _s) do
-        # elixir uses guard context for error messages
-        :elixir_rewrite.guard_rewrite(receiver, dot_meta, right, meta, e_args, "guard")
-      end
-    end
-
-    defp do_rewrite(_, receiver, dot_meta, right, meta, e_args, _s) do
-      {:ok, :elixir_rewrite.rewrite(receiver, dot_meta, right, meta, e_args)}
     end
   end
 end
